+++ src/main/client.ts
@@ -1,12 +1,17 @@
 import path from 'path'
 import { exec } from 'child_process'
-import { app } from 'electron'
+import { app, ipcMain } from 'electron'
 import * as settings from './settings'
 import * as php from './php'
 import { fileURLToPath } from 'url'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
+
+export const init = async () => {
+  ipcMain.on('client.execute', execute)
+  ipcMain.on('client.info', info)
+}
 
 function getClient() {
   const phpVersion = php.getVersion(settings.getSettings().php)
