+++ src/main/main.ts
@@ -1,16 +1,16 @@
-import { app, BrowserWindow, ipcMain, Tray, Menu, nativeImage } from 'electron'
+import { app, BrowserWindow, ipcMain } from 'electron'
 import { fileURLToPath } from 'url'
 import path from 'path'
+import log from 'electron-log/main'
+import * as dotenv from 'dotenv'
 import * as source from './source'
 import * as client from './client'
 import * as settings from './settings'
-import * as php from './php'
 import * as lsp from './lsp/index'
 import * as laravel from './laravel'
 import * as updater from './updater'
 import * as link from './link'
-import log from 'electron-log/main'
-import * as dotenv from 'dotenv'
+import * as tray from './tray'
 
 Object.assign(console, log.functions)
 
@@ -22,26 +22,6 @@ dotenv.config()
 export let window: BrowserWindow
 
 app.whenReady().then(async () => {
-  await lsp.init()
-
-  const trayIconPath: string = path.join(app.getAppPath(), 'src/main/tray/IconTemplate.png')
-  const tray = new Tray(nativeImage.createFromPath(trayIconPath))
-  const contextMenu = Menu.buildFromTemplate([
-    {
-      label: 'Check for updates',
-      click: async () => {
-        await updater.checkForUpdates()
-      },
-    },
-    {
-      label: 'Quit',
-      click: () => {
-        app.quit()
-      },
-    },
-  ])
-  tray.setContextMenu(contextMenu)
-
   window = new BrowserWindow({
     minWidth: 1100,
     minHeight: 700,
@@ -58,16 +38,12 @@ app.whenReady().then(async () => {
     center: true,
     icon: path.join(app.getAppPath(), 'build/icon.png'),
   })
-
   if (process.env.VITE_DEV_SERVER_URL) {
     await window.loadURL(process.env.VITE_DEV_SERVER_URL)
     window.webContents.openDevTools()
   } else {
     await window.loadFile(path.join(__dirname, `./index.html`))
   }
-
-  await updater.init()
-  await link.init()
 
   ipcMain.on('init', async event => {
     await laravel.init()
@@ -76,12 +52,14 @@ app.whenReady().then(async () => {
       settings: settings.getSettings(),
     })
   })
-  ipcMain.on('settings.store', settings.storeSettings)
-  ipcMain.on('client.execute', client.execute)
-  ipcMain.on('client.info', client.info)
-  ipcMain.on('source.open', source.open)
-  ipcMain.on('php.path', php.path)
-  // ipcMain.on('ssh.connect', ssh.connect)
+
+  await tray.init()
+  await lsp.init()
+  await updater.init()
+  await link.init()
+  await settings.init()
+  await client.init()
+  await source.init()
 })
 
 app.on('window-all-closed', () => {
