+++ src/main/source.ts
@@ -1,5 +1,9 @@
-import { dialog } from 'electron'
+import { dialog, ipcMain } from 'electron'
 import { IpcMainEvent } from 'electron'
+
+export const init = async () => {
+  ipcMain.on('source.open', open)
+}
 
 export const open = async (event: IpcMainEvent) => {
   // Use shell.openPath to open the specified folder
